requirements.txt
PyPDF2
python-dotenv
llama_index
llama_index.llms.ollama
llama_index.embeddings.huggingface
pycryptodome